2017年7月3日 1、实现HiveDaod的delete 操作 2、苏研平台不支持行级别操作。 2017年7月4日 1、已测试hive delete分区功能 2、使用url中principal 是hive/host@reaml，但是keytable 是user的。 2017年7月5日 1、优化hive delete 功能，可以删除之前的分区数据，但仅限义“partitionName”xml中参数开头，并以数字型区分的各个分区名。 2、优化publisher中 发布到hbase中直接传入token。

+++ DataController/src/main/java/com/eastcom/datacontroller/service/HiveServiceImpl.java
@@ -9,7 +9,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.io.IOException;
+import java.util.List;
 
 /**
  * Created by linghang.kong on 2017/7/4.
@@ -24,6 +26,9 @@ public class HiveServiceImpl implements HiveService<HiveJobs> {
 
     @Autowired
     private HiveEntity hiveEntity;
+
+    @Resource(name = "partitionName")
+    private String partitionName;
 
     @Override
     public String getName() {
@@ -43,13 +48,36 @@ public class HiveServiceImpl implements HiveService<HiveJobs> {
             hiveEntity.setTableName(table);
             hiveEntity.setPartition(entity.getPartition());
             hiveEntity.setSql(entity.getSql());
-            hiveDao.delete(hiveEntity);
+            if (!entity.isDf()) {
+                hiveDao.delete(hiveEntity);
+            }else {
+                // 仅用于删除时间分区
+                List<String> partitions = hiveDao.getPartitions(hiveEntity);
+                if (partitions !=null){
+                    int initPartition = subPartitionTime(hiveEntity.getPartition());
+                    for (String partition: partitions
+                         ) {
+                        if (subPartitionTime(partition) <= initPartition){
+                            hiveEntity.setPartition(partition);
+                            hiveDao.delete(hiveEntity);
+                        }
+                    }
+                }
+            }
         }
     }
 
     @Override
     public void close() throws IOException {
 
+    }
+
+
+    private int subPartitionTime(String partition){
+        if (partition !=null) {
+            return Integer.parseInt(partition.replaceAll(partitionName, ""));
+        }
+        return 0;
     }
